可以使用shell和pipe进行交互

+++ lib/spawn.c
@@ -303,7 +303,7 @@ copy_shared_pages(envid_t child)
 {
 	// LAB 5: Your code here.
 	int r;
-	for (uintptr_t addr = 0; addr < USTACKTOP; addr += PGSIZE){
+	for (uint32_t addr = 0; addr < USTACKTOP; addr += PGSIZE){
 		if((uvpd[PDX(addr)]&PTE_P)&&(uvpt[PGNUM(addr)]&PTE_P)&&(uvpt[PGNUM(addr)]&PTE_U)&&(uvpt[PGNUM(addr)]&PTE_SHARE)){
 			int perm=uvpt[PGNUM(addr)]&PTE_SYSCALL;
 			if((r=sys_page_map(0,(void*)addr,child,(void*)addr,perm))<0){
